fix: polish code
