+++ agents/generate/page-detail/analyze-page-detail.mjs
@@ -1,6 +1,7 @@
 import { access, readFile } from "node:fs/promises";
 import { join } from "node:path";
 import { getFileName, hasSourceFilesChanged } from "../../../utils/utils.mjs";
+import { generateFieldConstraints } from "../../../utils/generate-helper.mjs";
 import checkDetailResult from "../../utils/check-detail-result.mjs";
 
 export default async function analyzePageDetail(input, options) {
@@ -101,45 +102,6 @@ export default async function analyzePageDetail(input, options) {
       detailGenerated,
     };
   }
-
-  // Generate field usage constraints from builtinComponentLibrary
-  const generateFieldConstraints = (componentLibrary) => {
-    if (!componentLibrary || !Array.isArray(componentLibrary)) {
-      return "";
-    }
-
-    // Extract atomic fields
-    const atomicFields = componentLibrary.filter((comp) => comp.type === "atomic");
-
-    // Extract composite field combinations
-    const compositeFields = componentLibrary.filter((comp) => comp.type === "composite");
-
-    // Build constraints text
-    let constraints = "";
-
-    // Atomic fields section
-    constraints += "<fields_information>\n";
-    atomicFields.forEach((item) => {
-      const { field, summary } = item;
-      constraints += `- \`${field}\`: ${summary}\n`;
-    });
-    constraints += "</fields_information>\n\n";
-
-    // Composite combinations section
-    constraints += "<allowed_field_combinations>\n";
-    compositeFields.forEach((item) => {
-      constraints += `- \`${JSON.stringify(item.fieldCombinations)}\`: - **${item.name}** ${item.summary}\n`;
-    });
-    constraints += "</allowed_field_combinations>\n\n";
-
-    constraints +=
-      "- You can refer to the information in <fields_information> to understand what each field defines\n";
-    constraints +=
-      "- Each section MUST strictly follow the field combinations listed in <allowed_field_combinations>\n";
-    constraints += "    - DO NOT use any other field combinations\n";
-
-    return constraints;
-  };
 
   const fieldConstraints = generateFieldConstraints(builtinComponentLibrary);
