+++ utils/generate-helper.mjs
@@ -390,3 +390,46 @@ export const getComponentLibraryData = (tmpDir) => {
     return { componentLibrary: [] };
   }
 };
+
+/**
+ * Generate field usage constraints from builtin component library
+ * @param {Array} componentLibrary - Array of component definitions
+ * @returns {string} Formatted constraints text for LLM usage
+ */
+export function generateFieldConstraints(componentLibrary) {
+  if (!componentLibrary || !Array.isArray(componentLibrary)) {
+    return "";
+  }
+
+  // Extract atomic fields
+  const atomicFields = componentLibrary.filter((comp) => comp.type === "atomic");
+
+  // Extract composite field combinations
+  const compositeFields = componentLibrary.filter((comp) => comp.type === "composite");
+
+  // Build constraints text
+  let constraints = "";
+
+  // Atomic fields section
+  constraints += "<fields_information>\n";
+  atomicFields.forEach((item) => {
+    const { field, summary } = item;
+    constraints += `- \`${field}\`: ${summary}\n`;
+  });
+  constraints += "</fields_information>\n\n";
+
+  // Composite combinations section
+  constraints += "<allowed_field_combinations>\n";
+  compositeFields.forEach((item) => {
+    constraints += `- \`${JSON.stringify(item.fieldCombinations)}\`: - **${item.name}** ${item.summary}\n`;
+  });
+  constraints += "</allowed_field_combinations>\n\n";
+
+  constraints +=
+    "- You can refer to the information in <fields_information> to understand what each field defines\n";
+  constraints +=
+    "- Each section MUST strictly follow the field combinations listed in <allowed_field_combinations>\n";
+  constraints += "    - DO NOT use any other field combinations\n";
+
+  return constraints;
+}
